% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CI_MS95.R
\docType{data}
\name{CI_MS95}
\alias{CI_MS95}
\title{CI chondrite compositions (McDonough & Sun, 1995)}
\format{
dataframe
}
\usage{
CI_MS95
}
\description{
CI chondrite compositions (McDonough & Sun, 1995)
}
\references{
\url{https://doi.org/10.1016/0009-2541(94)00140-4}
}
\keyword{datasets}
