% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NMORB_SM89.R
\docType{data}
\name{NMORB_SM89}
\alias{NMORB_SM89}
\title{N-MORB compositions (Sun & McDonough, 1989)}
\format{
dataframe
}
\usage{
NMORB_SM89
}
\description{
N-MORB compositions (Sun & McDonough, 1989)
}
\references{
\url{https://doi.org/10.1144/GSL.SP.1989.042.01.19}
}
\keyword{datasets}
