% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM_MS95.R
\docType{data}
\name{PM_MS95}
\alias{PM_MS95}
\title{Primitive mantle compositions (McDonough & Sun, 1995)}
\format{
dataframe
}
\usage{
PM_MS95
}
\description{
Primitive mantle compositions (McDonough & Sun, 1995)
}
\references{
\url{https://doi.org/10.1016/0009-2541(94)00140-4}
}
\keyword{datasets}
