% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdg_filter_make.R
\name{spdg_filter_make}
\alias{spdg_filter_make}
\title{Defining filters for plotting spider diagrams}
\usage{
spdg_filter_make(
  Sample,
  Phase,
  Comment1,
  Comment2,
  Sample_value = NA,
  Phase_value = NA,
  Comment1_value = NA,
  Comment2_value = NA
)
}
\arguments{
\item{Sample}{(dataframe; chr) column for rock sample names}

\item{Phase}{(dataframe; chr) column for mineral names}

\item{Comment1}{(dataframe; chr) column for miscellaneous comments (1)}

\item{Comment2}{(dataframe; chr) column for miscellaneous comments (2)}

\item{Sample_value}{(chr) rock sample name you want to choose}

\item{Phase_value}{(chr) mineral name you want to choose}

\item{Comment1_value}{(chr) comment you want to choose (1)}

\item{Comment2_value}{(chr) comment you want to choose (2)}
}
\value{
List of 2: (1) all the row indexes under which data passing the filters (as a num vector) and (2) filter description (as a text)
}
\description{
\code{spdg_filter_make} defines filters for plotting spider diagrams.
You can set at most 4 different factors for the filtering.
This function can generate a filter with which all the conditions of
Sample = Sample_value, Phase = Phase_value, Comment1 = Comment1_value,
Comment2 = Comment2_value are satisfied. (If you set each 'xxx_value' as NA,
filtering will not be applied.)
}
\examples{
data <- spdgmake_testdata
Sample <- as.data.frame(data[,"Type"])
Phase <- as.data.frame(data[,"Type"])
Comment1 <- as.data.frame(data[,"Type"])
Comment2 <- as.data.frame(data[,"Type"])
filtering <- spdg_filter_make(Sample, Phase, Comment1, Comment2, NA, NA, NA, NA)


}
