% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdg_plot.R
\name{spdg_plot}
\alias{spdg_plot}
\title{Generating the plot area for spider diagrams}
\usage{
spdg_plot(PlotElements, Order_min = -3, Order_max = 3, Xlabelsize = 0.85)
}
\arguments{
\item{PlotElements}{(vector; chr) elements for which you want to plot concentrations}

\item{Order_min}{(int) log10(minimum value)}

\item{Order_max}{(int) log10(maximum value)}

\item{Xlabelsize}{(num) size of x-axis labels}
}
\description{
\code{spdg_plot} generates the plot area for spider diagrams.
}
\examples{
PlotElements <- c(
  "Cs", "Rb","Ba", "Pb", "Sr","La","Nd","Sm","Eu","Gd",
  "Dy","Yb","Lu","Hf","Ta","Y","Zr","Nb","Th","U"
  )
spdg_plot(PlotElements, -3, 3, 0.85)

}
