% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdgmake_testdata.R
\docType{data}
\name{spdgmake_testdata}
\alias{spdgmake_testdata}
\title{Testing data for spider diagram plotting (GLOSS)}
\format{
dataframe
}
\usage{
spdgmake_testdata
}
\description{
Testing data for spider diagram plotting (GLOSS)
}
\references{
\url{https://doi.org/10.1016/S0009-2541(97)00150-2}
}
\keyword{datasets}
