% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdg_filter_make.R
\name{spdg_filter_make}
\alias{spdg_filter_make}
\title{Defining filters for plotting spider diagrams}
\usage{
spdg_filter_make(
  Sample,
  Phase,
  Comment1,
  Comment2,
  Sample_value = NA,
  Phase_value = NA,
  Comment1_value = NA,
  Comment2_value = NA
)
}
\arguments{
\item{Sample}{Sample}

\item{Phase}{Sample}

\item{Comment1}{Sample}

\item{Comment2}{Sample}

\item{Sample_value}{Sample}

\item{Phase_value}{Sample}

\item{Comment1_value}{Sample}

\item{Comment2_value}{Sample}
}
\value{
List of 2: 1. all the row indexes under which data passing the filters (as a num vector); 2. filter description (as a text)
}
\description{
\code{spdg_filter_make} defines filters for plotting spider diagrams
}
\examples{
library(spdgmake)
data <- spdgmake_testdata
Sample <- as.data.frame(data[,"Type"])
Phase <- as.data.frame(data[,"Type"])
Comment1 <- as.data.frame(data[,"Type"])
Comment2 <- as.data.frame(data[,"Type"])
filtering <- spdg_filter_make(Sample, Phase, Comment1, Comment2, NA, NA, NA, NA)


}
