% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdg_plot.R
\name{spdg_plot}
\alias{spdg_plot}
\title{Generating the plot area for spider diagrams}
\usage{
spdg_plot(PlotElements, Order_min = -3, Order_max = 3, Xlabelsize = 0.85)
}
\arguments{
\item{PlotElements}{Sample}

\item{Order_min}{Sample}

\item{Order_max}{Sample}

\item{Xlabelsize}{Sample}
}
\description{
\code{spdg_plot} generates the plot area for spider diagrams
}
\examples{
library(spdgmake)
PlotElements <- c(
  "Cs", "Rb","Ba", "Pb", "Sr","La","Nd","Sm","Eu","Gd",
  "Dy","Yb","Lu","Hf","Ta","Y","Zr","Nb","Th","U"
  )
spdg_plot(PlotElements, -3, 3, 0.85)

}
