% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdg_plotdata_export.R
\name{spdg_plotdata_export}
\alias{spdg_plotdata_export}
\title{Exporting normalized data}
\usage{
spdg_plotdata_export(
  PlotElements,
  DataElements,
  Concentrations,
  Filt,
  Nml,
  Output = FALSE
)
}
\arguments{
\item{PlotElements}{Sample}

\item{DataElements}{Sample}

\item{Concentrations}{Sample}

\item{Filt}{Sample}

\item{Nml}{Sample}

\item{Output}{Sample}
}
\value{
Normalized concentration data
}
\description{
\code{spdg_plotdata_export} exports normalized data
}
\examples{
library(spdgmake)
data <- spdgmake_testdata

Sample <- as.data.frame(data[,"Type"])
Phase <- as.data.frame(data[,"Type"])
Comment1 <- as.data.frame(data[,"Type"])
Comment2 <- as.data.frame(data[,"Type"])

DataElements <- colnames(data)[4:ncol(data)]

Concentrations <- data[,4:ncol(data)]

PlotElements <- c(
  "Cs", "Rb","Ba", "Pb", "Sr","La","Nd","Sm","Eu","Gd",
  "Dy","Yb","Lu","Hf","Ta","Y","Zr","Nb","Th","U"
  )

spdg_plot(PlotElements, -3, 3, 0.85)

filtering <- spdg_filter_make(Sample, Phase, Comment1, Comment2, NA, NA, NA, NA)

spdg_lines(PlotElements, DataElements, Concentrations, filtering, NMORB_SM89, "black")

ProcessedData <- spdg_plotdata_export(PlotElements, DataElements, Concentrations, filtering, NMORB_SM89, Output = FALSE)

}
