% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spdg_plotdata_export.R
\name{spdg_plotdata_export}
\alias{spdg_plotdata_export}
\title{Exporting normalized data}
\usage{
spdg_plotdata_export(
  PlotElements,
  DataElements,
  Concentrations,
  Filt,
  Nml,
  Output = FALSE
)
}
\arguments{
\item{PlotElements}{(vector; chr) elements for which you want to plot concentrations}

\item{DataElements}{(vector; chr) elements presented in the input dataset}

\item{Concentrations}{(dataframe; num) concentrations (in µg/g or ppm) in the input dataset.
Note that each value in the dataframe should be a numeric type value.
NA values can be read but will not appear on the plot.}

\item{Filt}{(List of 2) filter you can generate by using spdg_filter_make().}

\item{Nml}{(dataframe; num) normalized factors. This package allows you to use
either of PM_MS95' (primitive mantle; McDonough & Sun 1995), 'CI_MS95'
(CI chondrite; McDonough & Sun 1995), and 'NMORB_SM89' (N-MORB; Sun & McDonough 1989).
However, you can set any normalized factors by defining a new 1 x n dataframe.}

\item{Output}{(logical) if Output = TRUE, normalized concentration data will be
exported as a .txt file in the working directory.}
}
\value{
Normalized concentration data
}
\description{
\code{spdg_plotdata_export} exports normalized concentration data.
}
\examples{
data <- spdgmake_testdata

Sample <- as.data.frame(data[,"Type"])
Phase <- as.data.frame(data[,"Type"])
Comment1 <- as.data.frame(data[,"Type"])
Comment2 <- as.data.frame(data[,"Type"])

DataElements <- colnames(data)[4:ncol(data)]

Concentrations_original <- data[,4:ncol(data)]
Concentrations <- as.data.frame(apply(Concentrations_original, c(1,2), as.numeric))

PlotElements <- c(
  "Cs", "Rb","Ba", "Pb", "Sr","La","Nd","Sm","Eu","Gd",
  "Dy","Yb","Lu","Hf","Ta","Y","Zr","Nb","Th","U"
  )

spdg_plot(PlotElements, -3, 3, 0.85)

filtering <- spdg_filter_make(Sample, Phase, Comment1, Comment2, NA, NA, NA, NA)

spdg_lines(PlotElements, DataElements, Concentrations, filtering, NMORB_SM89, "black")

ProcessedData <- spdg_plotdata_export(
  PlotElements, DataElements, Concentrations,
  filtering, NMORB_SM89, Output = FALSE
  )

}
